CPM-1235 Confirm delete note

+++ app/Http/Controllers/NotesController.php
@@ -130,15 +130,15 @@ class NotesController extends Controller
                 ->where('inbound_cpm_id', '=', $patientId)
                 ->where('outbound_cpm_id', '=', $author_id)
                 ->select(
-                                           [
-                                               'id',
-                                               'type',
-                                               'sub_type',
-                                               'attempt_note',
-                                               'scheduled_date',
-                                               'window_start',
-                                               'window_end',
-                                           ]
+                    [
+                        'id',
+                        'type',
+                        'sub_type',
+                        'attempt_note',
+                        'scheduled_date',
+                        'window_start',
+                        'window_end',
+                    ]
                                        )
                 ->get();
         }
@@ -202,6 +202,10 @@ class NotesController extends Controller
         $note = Note::findOrFail($noteId);
         if (Note::STATUS_DRAFT !== $note->status) {
             throw new \Exception('You cannot delete a non-draft note');
+        }
+
+        if ($note->author_id != auth()->id()) {
+            throw new \Exception('Only the author of the note can delete it');
         }
 
         $note->delete();
@@ -766,10 +770,10 @@ class NotesController extends Controller
     {
         return Practice::whereId($patient->program_id)
             ->where(
-                           function ($q) {
-                               $q->where('name', '=', 'phoenix-heart')
-                                   ->orWhere('name', '=', 'demo');
-                           }
+                function ($q) {
+                    $q->where('name', '=', 'phoenix-heart')
+                        ->orWhere('name', '=', 'demo');
+                }
                        )
             ->exists();
     }
